refactor(atsboard): extract the board recogniser into a shared package
RecognizeBoard, the atsBoards host table and the extraction modes were
service-private to internal/contribution, yet boardresolve already reached
across for them and link resolution is about to need them too. They move to
internal/atsboard as atsboard.Recognize (the package name already says
'board'), leaving contribution with the Greenhouse/Ashby job-id parsing that
really is service logic.

Behaviour is unchanged; the table's tests move with it. boardresolve now
depends on atsboard instead of contribution, so the cycle that would have
appeared once linksource needed the recogniser cannot form.
